\name{BSEHyrdoModels-package}
\alias{BSEHyrdoModels-package}
\alias{BSEHyrdoModels}
\docType{package}
\title{"BSEHyrdoModels"
\packageTitle{BSEHyrdoModels}
}
\description{
\packageDescription{BSEHyrdoModels}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{BSEHyrdoModels}
\packageIndices{BSEHyrdoModels}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{BSEHyrdoModels}

Maintainer: \packageMaintainer{BSEHyrdoModels}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
